+++ src/main/java/shadows/apotheosis/adventure/commands/GemCommand.java
@@ -56,7 +56,7 @@ public class GemCommand {
 			return 0;
 		})))))).then(Commands.literal("random").executes(c -> {
 			Player p = c.getSource().getPlayerOrException();
-			ItemStack gem = GemManager.getRandomGemStack(p.random, p.getLuck(), (ServerLevel) p.level);
+			ItemStack gem = GemManager.getRandomGemStack(p.random, p, (ServerLevel) p.level);
 			p.addItem(gem);
 			return 0;
 		})));